requirements.txt
python-frontmatter
markdown
beautifulsoup4